Refactored ccdb indexing in preparation of taxonomy
